pyproject.toml
[tool.poetry]
name = "prompt-hacker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "prompt_hacker"}]

[tool.poetry.dependencies]  
python = "^3.11"
requests = "^2.31.0"
openai = "^1.6.1"
ipykernel = "^6.28.0"
numpy = "^1.26.2"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
scipy = "^1.11.4"
pandas-stubs = "^2.1.4.231227"
types-tqdm = "^4.66.0.20240106"


[tool.poetry.group.dev.dependencies]
notebook = "^7.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
